+++ CHANGELOG.md
@@ -14,7 +14,10 @@
     * All:
       * Inherit from Badge (do not use this class directly)
       * With get\_badges\_for\_section(api, section, options={}) method
-      * With get\_badge\_data\_for\_section(api, section, badge\_key, term=nil, options={}) method
+      * With get\_badge\_data\_for\_section(api, section, badge, term=nil, options={}) method
+  * Addition of Badge::Requirements class
+  * Addition of Badge::Data class
+    * With update(api) method
 
 ## Version 0.2.2